+++ doc/TODO.md
@@ -1,0 +1,10 @@
+# To Do
+
+## Description
+
+A brief list of things to do for the project.
+
+## Tasks to do
+
+- Refactor any components that don't need to be full on class components
+  - similarly, decide if they should be a Pure Component, Stateless Component, or a Functional Component
