+++ src/react/components/base/Paragraph/Paragraph.tsx
@@ -1,10 +1,8 @@
-import React ,{ Component, ReactNode, ReactChild } from 'react';
+import React ,{ ReactElement } from 'react';
 
 import './Paragraph.css'
 
-// TODO Revise to be a Functional component
-
-export const Paragraph = (props: Local.ParagraphProps): JSX.Element => {
+export const Paragraph = (props: Local.ParagraphProps): ReactElement<any, any> => {
   const { children, type } = props;
   if (!children) throw new Error('Erorr expecting children elements');
   return (
