+++ src/stories/base.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 
 import { Button } from '../react/components/base/Button/Button';
-import { Field } from '../react/components/base/Field/Field';
+import { Field } from '../react/components/stateful/Forms/Field/Field';
 import { Heading } from '../react/components/base/Heading/Heading';
 import { Paragraph } from '../react/components/base/Paragraph/Paragraph';
